Move related types
