+++ src/pages/registration/resource_type_tab/sub_type_forms/research_data_types/DatasetForm.tsx
@@ -6,10 +6,8 @@ import { useTranslation } from 'react-i18next';
 import { FetchResultsParams, fetchResults } from '../../../../../api/searchApi';
 import { EmphasizeSubstring } from '../../../../../components/EmphasizeSubstring';
 import { ResearchDataType, ResourceFieldNames } from '../../../../../types/publicationFieldNames';
-import {
-  ConfirmedDocument,
-  ResearchDataRegistration,
-} from '../../../../../types/publication_types/researchDataRegistration.types';
+import { ResearchDataRegistration } from '../../../../../types/publication_types/researchDataRegistration.types';
+import { ConfirmedDocument } from '../../../../../types/registration.types';
 import { dataTestId } from '../../../../../utils/dataTestIds';
 import { useDebounce } from '../../../../../utils/hooks/useDebounce';
 import { findRelatedDocumentIndex, getTitleString } from '../../../../../utils/registration-helpers';
